Update docs, add comments, reorganize code into smaller files.
